requirements.txt
openpyxl>=3.1,<4
httpx>=0.27,<1
pydantic>=2.8,<3
schwab-py>=1.5.1,<2
pytest>=8.3,<9
